requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.2.1
python-telegram-bot==21.6
httpx==0.27.2
tenacity==9.0.0
